Break the Article Form Field into its own reusable component
The Form will be reused in the Update Article feature in addition to the Create Article.
