--- src/app/modules/homepage/create-article-form/index.tsx
@@ -1,5 +1,0 @@
-export default function CreateArticleForm() {
-  return (
-    <div></div>
-  )
-}
